refactor: correctly use tslint to fix
