+++ src/rulesets/oas/__tests__/info-license.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../index.json';
+import * as ruleset from '../rules.json';
 
 describe('info-license', () => {
   const s = new Spectral();
