+++ src/rulesets/oas/__tests__/no-script-tags-in-markdown.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../index.json';
+import * as ruleset from '../rules.json';
 
 describe('no-script-tags-in-markdown', () => {
   const s = new Spectral();
