+++ src/rulesets/oas/__tests__/openapi-tags-alphabetical.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../index.json';
+import * as ruleset from '../rules.json';
 
 describe('openapi-tags-alphabetical', () => {
   const s = new Spectral();
