+++ src/rulesets/oas/__tests__/path-declarations-must-exist.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../index.json';
+import * as ruleset from '../rules.json';
 
 describe('path-declarations-must-exist', () => {
   const s = new Spectral();
