+++ src/rulesets/oas/__tests__/path-keys-no-trailing-slash.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../index.json';
+import * as ruleset from '../rules.json';
 
 describe('path-keys-no-trailing-slash', () => {
   const s = new Spectral();
