+++ src/rulesets/oas/functions/oasPathParam/__tests__/index.test.ts
@@ -1,7 +1,7 @@
 import { RuleType, Spectral } from '../../../../../index';
 import { commonOasFunctions } from '../../../index';
 
-import { rules } from '../../../index.json';
+import { rules } from '../../../rules.json';
 
 const ruleset = { functions: commonOasFunctions(), rules };
