+++ src/rulesets/oas2/__tests__/oas2-schema.ts
@@ -1,5 +1,5 @@
 import { RuleType, Spectral } from '../../../spectral';
-import * as ruleset from '../index.json';
+import * as ruleset from '../rules.json';
 
 describe('oas2-schema', () => {
   const s = new Spectral();
